separated examples for image and titles
