+++ Example/JKSteppedProgressBar/ViewController.swift
@@ -13,24 +13,12 @@ class ViewController: UIViewController {
 
     @IBOutlet weak var currentTabLabel: UILabel!
     @IBOutlet weak var progressbar: SteppedProgressBar!
-    @IBOutlet weak var progressbarWithImages: SteppedProgressBar!
     
     @IBOutlet weak var nextButton: UIButton!
     @IBOutlet weak var prevButton: UIButton!
     
     var inset: UIEdgeInsets {
         return UIEdgeInsets(top: 10, left: 10, bottom: 30, right: 30)
-    }
-    
-    // MARK: UI configure methods
-    func configureProgressBarWithImages() {
-        progressbarWithImages.insets = inset
-        progressbarWithImages.titles = ["Image 1".localized, "Image 2".localized, "Image 3".localized,]
-        progressbarWithImages.images = [
-            UIImage(named: "DaisyDuck")!,
-            UIImage(named: "MickeyMouse")!,
-            UIImage(named: "MinnieMouse")!,
-        ]
     }
     
     func configureTitleProgressBar() {
@@ -43,7 +31,6 @@ class ViewController: UIViewController {
         nextButton.isEnabled = currentTab < progressbar.titles.count
         prevButton.isEnabled = currentTab > 0
         progressbar.currentTab = currentTab
-        progressbarWithImages.currentTab = currentTab
         currentTabLabel.text = "\(currentTab)"
     }
     
@@ -51,7 +38,6 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         configureTitleProgressBar()
-        configureProgressBarWithImages()
         updateButtons(0)
     }
